test(button): abstract test args to setup file
