+++ lib/components/button/button.a11y.test.ts
@@ -1,25 +1,21 @@
+import testArgs from "./button.test.setup";
 import { runA11yTests } from "../../test/a11y-test-utils";
 import "../../index";
 
 describe("button", () => {
     runA11yTests({
-        baseClass: "s-btn",
-        variants: ["danger", "muted"],
-        modifiers: {
-            primary: ["filled", "outlined"],
-            secondary: [...["xs", "sm", "md"], ...["dropdown", "icon"]],
-            global: ["is-loading"],
-            standalone: [
-                ...["link", "unset"],
-                ...["facebook", "github", "google"],
-            ],
-        },
-        attributes: {
-            type: "button",
-        },
-        children: {
-            default: "Ask question",
-        },
-        tag: "button",
+        ...testArgs,
+        excludedTestids: [
+            /^s-btn-(?=.*unset).*badge$/, // s-btn with badge and unset variant not supported
+        ],
+        skippedTestids: [
+            // TODO resolve btn badge contrast issues
+            // matches tests with a badge in light, dark, and highcontrast-light modes
+            /s-btn-(light|dark).*?badge/,
+            // matches tests with a badge in highcontrast-light modes, excluding filled, danger, github, facebook, sm, or xs
+            /s-btn-highcontrast-light-(?!.*(filled|danger|github|facebook|sm|xs)).*?badge/,
+            // matches tests with a badge in highcontrast-light modes, are muted and/or outlined, and are sm or xs
+            /s-btn-highcontrast-light-(?:muted-outlined-|muted-|outlined-)?(?:sm|xs).*?badge/,
+        ],
     });
 });
