+++ lib/components/button/button.visual.test.ts
@@ -1,41 +1,11 @@
-import { html } from "@open-wc/testing";
+import testArgs from "./button.test.setup";
 import { runVisualTests } from "../../test/visual-test-utils";
 import "../../index";
-
-const getChild = (child?: string): string => {
-    switch (child) {
-        case "badge":
-            return `Ask question
-                <span class="s-btn--badge">
-                    <span class="s-btn--number">198</span>
-                </span>`;
-        default:
-            return "Ask question";
-    }
-};
+import { html } from "@open-wc/testing";
 
 describe("button", () => {
-    // TODO test disabled states, interaction pseudo-classes
     runVisualTests({
-        baseClass: "s-btn",
-        variants: ["danger", "muted"],
-        modifiers: {
-            primary: ["filled", "outlined"],
-            secondary: [...["xs", "sm", "md"], ...["dropdown", "icon"]],
-            global: ["is-loading"],
-            standalone: [
-                ...["link", "unset"],
-                ...["facebook", "github", "google"],
-            ],
-        },
-        attributes: {
-            type: "button",
-        },
-        children: {
-            default: getChild(),
-            badge: getChild("badge"),
-        },
-        tag: "button",
+        ...testArgs,
         template: ({ component, testid }) => html`
             <div
                 class="bg-black-225 d-inline-flex ai-center jc-center hs1 ws2 p8"
@@ -43,6 +13,6 @@ describe("button", () => {
             >
                 ${component}
             </div>
-        `,
+        `
     });
 });
