+++ lib/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
     "extends": "./tsconfig.json",
-    "exclude": ["test/**/*.ts", "**/*.test.ts"]
+    "exclude": ["test/**/*.ts", "**/*.test.ts", "**/*.setup.ts"]
 }
